domains/manufacturing/mcp_server/requirements_agentcore.txt
# Requirements for AgentCore deployment
bedrock-agentcore
boto3>=1.34.0
botocore>=1.34.0

# MCP Framework
fastmcp==0.1.0
uvicorn[standard]==0.24.0

# Optimization Solvers
ortools==9.8.3296
pulp==2.7.0
cvxpy==1.4.1
osqp==0.6.2
clarabel==0.6.0
qpsolvers==3.4.0

# Mathematical Libraries
numpy==1.24.3
scipy==1.11.1
pandas==2.0.3

# Data Validation
pydantic==2.4.2

# HTTP and API
httpx==0.24.1
requests==2.31.0

# JSON and Data Processing
orjson==3.9.4

# Environment and Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Time and Date
python-dateutil==2.8.2
pytz==2023.3
